Naprawione wyświetlanie newsów
Naprawione wyświetlanie newsów przeniesienie klasy RecyclerAdapter do
Fragments\News.cs
